+++ datadog_checks_dev/datadog_checks/dev/tooling/constants.py
@@ -12,6 +12,12 @@ INTEGRATION_REPOS = [
     'integrations-extras',
     'integrations-internal',
 ]
+REPO_CHOICES = {
+    'core': 'integrations-core',
+    'extras': 'integrations-extras',
+    'internal': 'integrations-internal',
+    'agent': 'datadog-agent',
+}
 VERSION_BUMP = {
     'Added': semver.bump_minor,
     'Changed': semver.bump_major,